move config base to __init__
