+++ agent_app_sample_code/cookbook/config/agents/genie_agent.py
@@ -1,5 +1,5 @@
 from typing import Any, List
-from cookbook.config.base import SerializableConfig
+from cookbook.config import SerializableConfig
 from mlflow.models.resources import DatabricksResource, DatabricksGenieSpace
