+++ agent_app_sample_code/cookbook/config/data_pipeline/data_pipeline_output.py
@@ -1,4 +1,4 @@
-from cookbook.config.base import SerializableConfig
+from cookbook.config import SerializableConfig
 from typing import Optional
 
 from databricks.sdk import WorkspaceClient
