+++ agent_app_sample_code/cookbook/config/shared/agent_storage_location.py
@@ -11,7 +11,7 @@ from pydantic import Field
 from databricks.sdk import WorkspaceClient
 from databricks.sdk.errors.platform import ResourceDoesNotExist
 from databricks.sdk.errors import NotFound
-from cookbook.config.base import SerializableConfig
+from cookbook.config import SerializableConfig
 from databricks.sdk import WorkspaceClient
